# FlatGFA: Optimize GFA parsing a bit