--- a/polbin/Cargo.toml
+++ b/polbin/Cargo.toml
@@ -10,3 +10,7 @@ gfa = "0.10.1"
 memmap = "0.7.0"
 num_enum = "0.7.2"
 zerocopy = { version = "0.7.32", features = ["derive"] }
+
+[profile.profiling]
+inherits = "release"
+debug = true